wifi: brcmfmac: Rename Cypress 89459 to BCM4355
The commit that introduced support for this chip incorrectly claimed it
is a Cypress-specific part, while in actuality it is just a variant of
BCM4355 silicon (as evidenced by the chip ID).

The relationship between Cypress products and Broadcom products isn't
entirely clear and employees from the companies haven't been helpful in
clarifying it, but given what little information is available and prior
art in the driver, it seems the convention should be that originally
Broadcom parts should retain the Broadcom name.

Thus, rename the relevant constants and firmware file. Also rename the
specific 89459 PCIe ID to BCM43596, which seems to be the *actual*
subvariant name for this PCI ID (as defined in the out-of-tree bcmdhd
driver).

Fixes: dce45ded7619 ("brcmfmac: Support 89459 pcie")
Signed-off-by: Hector Martin <marcan@marcan.st>