drm/mediatek: Remove all preprocessor ifs for CONFIG_MTK_CMDQ
Since this driver was migrated to use the MediaTek CMDQ helpers,
it's not anymore necessary to exclude CMDQ related code with
preprocessor if branches, as CMDQ is optional and the helpers
are providing the necessary inline functions to manage the case
in which CONFIG_MTK_CMDQ is not set.

Clean up all instances of `#if IS_REACHABLE(CONFIG_MTK_CMDQ)`
from all drivers in drm/mediatek.

Signed-off-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>