UBUNTU: SAUCE: apparmor4.0.0 [71/92]: apparmor: fix notification header size
BugLink: http://bugs.launchpad.net/bugs/2028253

When strings are appended to the notification the header size should
be updated to reflect the correct size. While the size is also
directly returned as part of delivering the notification, the header
should also be update to conform to specification and allow for
verification.

If verification is enabled and the notification contains appended
strings then notifications fail verification and won't be delivered.

Fixes: 9a3b87d8b9a0 ("UBUNTU: SAUCE: apparmor4.0.0 [71/92]: prompt - rework build to use append fn, to simplify adding strings")
Signed-off-by: John Johansen <john.johansen@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Roxana Nicolescu <roxana.nicolescu@canonical.com>
(cherry picked from commit 44bf9e7a92fadb32e0c4abb4438d9991acdae31e)
Signed-off-by: Paolo Pisati <paolo.pisati@canonical.com>