mesa-ci: migrate the s3 bucket to mesa-rootfs with 1y expiration time.
The bucket in use to upload artifacts to S3, mesa-lava, has an expiration
period of a month. If there isn't a kernel uprev in a month, the build job
should be retriggered to regenerate them. Meanwhile, some jobs can fail
simply because the expired artifacts. A year is a time window that ensures
a kernel uprev.

Signed-off-by: Sergi Blanch Torne <sergi.blanch.torne@collabora.com>