arm64: dts: mediatek: mt8195-cherry: Add a graph for display IPs
The display IPs in MT8195 are flexible and those support being
interconnected with different instances of DDP IPs and/or with
different DDP IPs, forming a full Display Data Path that ends
with an actual display output, which is board-specific.

Add a common graph in the main mt8195.dtsi DT, shared between
all of the currently supported boards (including IoT and EVK)
to avoid cluttering device-specific DTs, and add the relevant
board-specific graph in mt8195-cherry, shared between all of
the currently supported MT8195 Chromebooks.

Signed-off-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>