arm64: dts: mt8195: Add UFSHCI node and change UFSPHY clocks
Add a node for the Universal Flash Storage controller and keep it
disabled by default.
While at it, also change the UFS PHY node to use the right clocks
for unipro and mp to improve reliability on platforms that don't
enable, or that disable, UFS in the bootloader before booting the
kernel.

Signed-off-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>