Revert "soc: qcom: pd_mapper: Add X1E80100"
This reverts commit bd6db1f1486eedb8460647b3fbe9b5ae8fd09207.

Enabling the in-kernel pd-mapper means that services become available
sooner and sometimes even before the firmware is ready to receive
requests. This in turn breaks services like battery notifications,
altmode notifications and audio.

This is easily triggered on x1e80100, but this is not platform specific.

Revert for x1e80100 for now until the underlying issues have been
addressed.

Fixes: bd6db1f1486e ("soc: qcom: pd_mapper: Add X1E80100")
Link: https://lore.kernel.org/lkml/Zqet8iInnDhnxkT9@hovoldconsulting.com/
Signed-off-by: Johan Hovold <johan+linaro@kernel.org>