WIP: arm64: dts: qcom: x1e78100-t14s: Enable the USB2 controller and PHYs
The Thinkpad T14s seems to have the third PTN2333 redriver populated,
which, if the same design is the same as on QCP, it belongs to the USB2
controller. On QCP this is connected to the USB5 port, but on T14s,
without schematics we don't know yet what it is connected to.
Add the redriver and enable the eUSB2 PHY.

Not-signed-off-by: Abel Vesa <abel.vesa@linaro.org>