arm64: dts: qcom: x1e78100-qcp: Enable support for both Type-A USB ports
The Qualcomm X1E QCP has 2 USB-A ports that are connected to the USB
multiport controller via 2 eUSB2 redriver and PHY chains for high-speed
support and 2 QMP PHYs for superspeed support. Enable the MP controller
and all related PHYs and redrivers in order to make those 2 USB-A ports
to work.

Signed-off-by: Abel Vesa <abel.vesa@linaro.org>