HACK: MTK: Disable AFBC support
Starting with Mesa 25 AFBC will get more eagerly used. Unfortunately it
seems both primary and overlay planes on Genio 700 and Genio 1200 are
broken in this regard. On 1200 it results in odd corrupted frames, while
on 700 it results in vblank timeouts (for primary planes)

Add a HACK to disable AFBC for now until this can be further diagnosed