pmdomain: mediatek: mtk-scpsys: Remove duplicated MT8173 support
Remove support for the MediaTek MT8173 MTCMOS from this old and
deprecated power domains driver.

The MT8173 SoC was migrated to mtk-pm-domains a lot of time ago
(many kernel releases ago) because it was not working correctly
in this driver, so removing support from there is not breaking
anything - as support for this SoC was not realy functional in
this driver's code.

Signed-off-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>