soc: mediatek: Add MediaTek Secure world Power Management driver
Add a driver to communicate with the Secure world Power Management
MCU: though this is present on all MediaTek SoCs, in this initial
driver, only SSPM Version 3 time synchronization is provided.

Time synchronization is required for proper suspend/resume of the
various co-processors present in the SoC and, in some instances,
also required for bandwidth management and/or frequency scaling of
the co-processor.

Signed-off-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>