arm64: dts: mt8365-evk: Disable it66121_hdmi node by default
On the Mediatek Genio 350-EVK board, Ethernet and HDMI (DPI0) are
sharing pins, and only one functionality can be selected at a time with
a DIP switch. In the board devicetree, this leads to have dpi0 and
ethernet nodes disabled by default.
But the it66121_hdmi node for the ITE IT66121FN HDMI Bridge IC, present
on the board, is always enabled and it is not useful when dpi0 is
disabled, so disable it too by default.

Signed-off-by: Louis-Alexis Eyraud <louisalexis.eyraud@collabora.com>