UBUNTU: SAUCE: selftests: gpio: gpio-aggregator: add a test case for _sysfs prefix reservation
BugLink: https://bugs.launchpad.net/bugs/2103496

The kernel doc for gpio-aggregator configfs interface, which was recently
added, states that users should not be able to create an aggregator with a
name prefixed by "_sysfs" via configfs. However, it was found that this
guard does not function as expected (thanks to Dan Carpenter for
identifying and fixing the issue).

Add a test case to verify the guard.

Signed-off-by: Koichiro Den<koichiro.den@canonical.com>
Link:https://lore.kernel.org/r/20250412150119.1461023-1-koichiro.den@canonical.com
Signed-off-by: Bartosz Golaszewski<brgl@bgdev.pl>

(cherry picked from commit 290ffcfe3042dcd5b701a98edd48d7962d35556b gpio/for-next)
Signed-off-by: Koichiro Den<koichiro.den@canonical.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>