ci: keep all DTBs
These files are tiny, there is no reason to have to edit this file for
each new device that we want to support in CI; let's keep all the DTBs.

Note that with ~3k DTBs per build, with 50 builds per major release
being merged (which is way overestimated), it would take 80+ years to
reach the 50M limit on the number of objects in the bucket.