HACK: ASoC: SOF: core: Defer probe if firmware missing
As long as the deferred probe timeout hasn't expired, defer SOF probe if
the firmware isn't present. This takes care of the case where the
initramfs doesn't contain the firmware, so SOF can defer until the
rootfs with the firmware is loaded, and then still probe successfully.

That is the case with the existing debian ramdisk and NFS rootfs booted
by the Genio boards in LAVA, and having SOF probe gets the DT kselftest
passing.

This implementation assumes CONFIG_SND_SOC_SOF_PROBE_WORK_QUEUE=n, as
otherwise the firmware loading will happen in a work queue after the
driver's probe has completed, so this is meant as a stopgap solution
until a proper approach is developed.

Signed-off-by: Nícolas F. R. A. Prado <nfraprado@collabora.com>