ci: enable local builds, only call ci-fairy s3cp when needed
Defer the check on S3_HOST to when it's really needed, and only set the
relevant variables and call ci-fairy s3cp when S3_HOST is defined.

In particular this allows developers to build the kernel by calling
.gitlab-ci/build.sh locally and inspect the result without involving CI
at all.

E.g.:

$ KERNEL_ARCH="x86_64" ./.gitlab-ci/build.sh

This can speed up the development/test cycle, and also make it easier to
perform temporary/local experiments.