dt-bindings: power: mt8196-gpufreq: Describe nvmem provider ability
On the MediaTek MT8196 SoC, the Mali GPU's "shader_present" hardware
register does not correctly describe the present shader cores. Instead,
an efuse provides this information.

However, the efuse address is, to quote the vendor, "confidential", and
we are not allowed to know it. The information we do know, however, is
that the GPUEB MCU's firmware does read this efuse, and exposes the
value in the memory it purposefully shares with the system for accessing
such information.

We can therefore describe the mediatek,mt8196-gpufreq device as being an
nvmem provider for this purpose, in order to avoid infringing on the
vendor's hard-earned intellectual property (efuse address).

Signed-off-by: Nicolas Frattaroli <nicolas.frattaroli@collabora.com>