HACK: drm/mediatek: Add remove_driver_render module parameter for Panthor
The mediatek-drm driver should not have the DRIVER_RENDER feature,
because this is *not* a GPU driver - and having it is actually
wrong (and will create a number of issues).

However, for whatever reason, the Mali DDK needs that.

Let this driver register with either the proper drm_driver features
or the wrong one by adding and checking the "remove_driver_render"
module parameter, useful in case panthor is used.

Also, vary the defaults depending on whether Mali or Panthor are
included in the kernel build.

Signed-off-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>